third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/prefixed-header-deprecation-warning-expected.txt

CONSOLE WARNING: The 'X-WebKit-CSP' headers are deprecated; please consider using the canonical 'Content-Security-Policy' header instead.
CONSOLE ERROR: line 9: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*".

This test passes if it doesn't alert fail, and a deprecation warning is presented in the console.
